TUSCANY-1237 - Submitting Ron's patch.
git-svn-id: https://svn.apache.org/repos/asf/incubator/tuscany/java/das@557695 13f79535-47bb-0310-9956-ffa450edef68

+++ rdb/src/main/java/org/apache/tuscany/das/rdb/impl/InsertCommandImpl.java
@@ -57,7 +57,7 @@ public class InsertCommandImpl extends WriteCommandImpl {
 
     public int getGeneratedKey() {
         try {
-            return statement.getGeneratedKey();
+            return statement.getGeneratedKey().intValue();
         } catch (SQLException ex) {
             throw new RuntimeException(ex);
         }
